# openapi/v0.yaml
openapi: 3.1.0
info:
  title: Fluorite Registry API
  version: "0.1.0"
  description: >
    Registry for publishing and fetching compiled TurboWarp extensions built
    with the Fluorite Compiler. All endpoints are versioned under /v0.
servers:
  - url: https://fluorite.sdisk.us/v0
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Accepts two kinds of bearer token: session tokens (from /auth/login,
        full access, 7-day TTL) and automation tokens (from /auth/tokens,
        long-lived, restricted to scopes granted at creation — currently only
        "publish"). Scope enforcement happens server-side per-endpoint.
  parameters:
    namespaceParam:
      name: namespace
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$"
...
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    Error:
      description: Standard error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated, but not permitted to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VERSION_PENDING_REVIEW
            message:
              type: string
              example: >-
                You have another version pending review. It needs to be
                approved or rejected before you can publish again.
            field:
              type: [string, "null"]
    User:
      type: object
      required: [namespace, displayName, type, trusted]
      properties:
        namespace:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [admin, normal]
        trusted:
          type: boolean
          description: >-
            Whether this user's publishes are auto-approved (true after their
            first-ever publish has been reviewed and approved).
...
    Notification:
      type: object
      required: [id, message, createdAt, readAt]
      properties:
        id:
          type: string
        message:
          type: string
        packageId:
          type: [string, "null"]
        version:
          type: [string, "null"]
        readAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags: [Open]
      operationId: listUsers
      summary: List users.
      security: []
      parameters:
        - $ref: "#/components/parameters/cursorParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  nextCursor:
                    type: [string, "null"]
    post:
      tags: [Adminspace]
      operationId: createUser
      summary: Create an account directly, without self-signup.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Error"
  /users/{namespace}:
    get:
      tags: [Open]
      operationId: getUser
      summary: Look up a user.
      security: []
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Userspace]
      operationId: updateUser
      summary: Update your own profile. Admins may also update another user's.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: >-
            Updated. Changing password revokes all existing sessions AND
            automation tokens for this user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Userspace]
      operationId: deleteUser
      summary: Delete your own account, or (as admin) another user's, and everything they own.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      responses:
        "204":
          description: Deleted.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{namespace}/role:
    patch:
      tags: [Adminspace]
      operationId: updateUserRole
      summary: Promote or demote a user.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [admin, normal]
      responses:
        "200":
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{namespace}/trust:
    patch:
      tags: [Adminspace]
      operationId: updateUserTrust
      summary: Reset whether this user's publishes are auto-approved.
      parameters:
        - $ref: "#/components/parameters/namespaceParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trusted]
              properties:
                trusted:
                  type: boolean
      responses:
        "200":
          description: Trust flag updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /notifications:
    get:
      tags: [Userspace]
      operationId: listNotifications
      summary: List your own notifications.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [unread, read]
        - $ref: "#/components/parameters/cursorParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A page of notifications.
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  nextCursor:
                    type: [string, "null"]
    delete:
      tags: [Userspace]
      operationId: clearReadNotifications
      summary: Bulk-delete your own notifications by status (e.g. ?status=read).
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [read]
      responses:
        "204":
          description: Cleared.
  /notifications/{id}:
    patch:
      tags: [Userspace]
      operationId: markNotificationRead
      summary: Mark a notification read. Returns 404 (not 403) if it belongs to someone else.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Userspace]
      operationId: deleteNotification
      summary: Delete a single notification you own.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"